bigger package

+++ gms-server/src/main/java/org/gms/server/StorageInventory.java
@@ -33,11 +33,11 @@ import java.util.*;
 public class StorageInventory {
     private final Client c;
     private Map<Short, Item> inventory = new LinkedHashMap<>();
-    private final byte slotLimit;
+   private final short slotLimit;
 
     public StorageInventory(Client c, List<Item> toSort) {
         this.inventory = new LinkedHashMap<>();
-        this.slotLimit = (byte) toSort.size();
+       this.slotLimit = (short) toSort.size();
         this.c = c;
 
         for (Item item : toSort) {
@@ -45,7 +45,7 @@ public class StorageInventory {
         }
     }
 
-    private byte getSlotLimit() {
+    private short getSlotLimit()  {
         return slotLimit;
     }
